// src/api-serverless/src/generated/models/DropComment.ts
/**
 * Seize API
 * Following is the API interface description for most common operations in Seize API. Some modifying endpoints may need authentication token.
 *
 * OpenAPI spec version: 1.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { ProfileMin } from '../models/ProfileMin';
import { HttpFile } from '../http/http';

export class DropComment {
  'id': number;
  'author': ProfileMin;
  'comment': string;
  'created_at': number;

  static readonly discriminator: string | undefined = undefined;

  static readonly attributeTypeMap: Array<{
    name: string;
    baseName: string;
    type: string;
    format: string;
  }> = [
    {
      name: 'id',
      baseName: 'id',
      type: 'number',
      format: 'int64'
    },
    {
      name: 'author',
      baseName: 'author',
      type: 'ProfileMin',
      format: ''
    },
    {
      name: 'comment',
      baseName: 'comment',
      type: 'string',
      format: ''
    },
    {
      name: 'created_at',
      baseName: 'created_at',
      type: 'number',
      format: 'int64'
    }
  ];

  static getAttributeTypeMap() {
    return DropComment.attributeTypeMap;
  }

  public constructor() {}
}